refactor(libsarga): single-home tty/auth readers and cfg-gate allocator machinery to the sarga target
Consolidate terminal echo/read_line/read_password into libsarga::tty and
the auth cap constants into libsarga::auth, adding the shared whole-file
readers (io::read_to_end, read_stdin_all_bytes) the login/passwd/su
consumers previously duplicated. The slab allocator, global allocator
static and memcpy/memset/memcmp/memmove runtime shims are now
cfg(target_os = "none")-gated like the panic/alloc_error handlers: on the
host they were dead code and, on new clippy, suspicious runtime-symbol
definitions. Also drop the useless `n as usize` casts in the getdents
loop and write SHEBANG as the byte str `*b"#!"` (clippy fixes).

+++ libsarga/src/libskyos.rs
@@ -67,8 +67,8 @@ pub fn list_dir(path: &str) -> Option<Vec<String>> {
             _ => break,
         };
         let mut off = 0;
-        while off < n as usize {
-            if off + 18 > n as usize {
+        while off < n {
+            if off + 18 > n {
                 break;
             }
             let d_ino = u64::from_ne_bytes(buf[off..off + 8].try_into().unwrap());
